+++ taichi/ir/ir.cpp
@@ -980,4 +980,12 @@ void Stmt::infer_type() {
     irpass::typecheck(this);
 }
 
+void Cache(int v, const Expr &var) {
+  dec.scratch_opt.push_back(std::make_pair(v, var.snode()));
+}
+
+void CacheL1(const Expr &var) {
+  dec.scratch_opt.push_back(std::make_pair(1, var.snode()));
+}
+
 TLANG_NAMESPACE_END
